new model and config file

+++ restapi/src/main/java/pl/lukasz/fd/restapi/Controller/GlobalRepositoryController.java
@@ -16,7 +16,7 @@ public class GlobalRepositoryController
 
     @GetMapping("/GetAccount")
     @ResponseBody
-    public List<Accounts> GetAccount(@RequestParam(name = "id") Long id)
+    public List<Accounts> GetAccount(@RequestParam(name = "id") Integer id)
     {
         return _globalRepositoryService.GetAccount(id);
     }
